add doc

+++ create-train.py
@@ -1,5 +1,28 @@
 #!/usr/bin/env python3
-
+"""
+Usage:      python create-train.py $model
+Motivation: Create the structure of data-train where the data for different experiments are saved.
+            net-mode, log, qlog: keep the information about the training
+basedir
+├── ...
+├── data-train               # Recognition files and links to important parts for epoch=1
+│   ├── qlog          		 # dir to collect qlog data
+│   ├── log         	     # dir to collect log data
+│   ├── net-model            # dir to save the checkpoints
+│   ├── train.q.sh           # script to start the training
+│   ├── base                 -> basedir
+│   ├── data-train           -> basedir/data-train
+│   ├── tools                -> basedir/tools
+│   ├── tools-recog          -> basedir/tools-recog
+│   ├── flow                 -> basedir/flow
+│   ├── features             -> basedir/features
+│   ├── features.warped      -> basedir/features.warped
+│   ├── config               -> basedir/config
+│   ├── config-train         -> basedir/config-train
+│   ├── dependencies         -> basedir/dependencies
+│   ├── sprint-executables   -> basedir/sprint-executables
+└──
+"""
 import faulthandler
 faulthandler.enable()
 import better_exchook
@@ -11,6 +34,7 @@ import sys
 import shutil
 from tools import Settings
 
+
 argparser = argparse.ArgumentParser()
 argparser.add_argument("model")
 args = argparser.parse_args()
@@ -19,6 +43,10 @@ mydir = Settings.base_dir
 print("Base dir:", mydir)
 os.chdir(mydir)
 
+# make sure the source of data-train link exists
+assert os.path.exists(os.readlink("data-train")), \
+    'Path data-train is a broken symlink. You have to call "setup-data-dir.py" to setup the symlink to working dir.'
+
 # Just take the base config filename as name for the training.
 # This also includes the epoch.
 model = args.model
@@ -26,21 +54,26 @@ config_file = "config-train/%s.config" % model
 assert os.path.exists(config_file)
 
 train_dir = "data-train/%s" % model
-print("Create %s" % train_dir)
+if os.path.exists(train_dir):
+    sys.exit("The directory %s exists already. We are stopping the script.." % train_dir)
 
+
+print("Create %s" % train_dir)
 os.mkdir(train_dir)
 os.mkdir(train_dir + "/qdir")
 os.mkdir(train_dir + "/log")
 os.mkdir(train_dir + "/net-model")
 
+# training script
 shutil.copy("train.q.sh", train_dir)
+# info file
 with open(train_dir + "/settings.sh", "w") as f:
     f.write("model=%s\n" % model)
     f.write("setup_basedir=%s\n" % mydir)
 
+# Symlinks
 os.symlink(mydir, train_dir + "/base")
 
-#test -e $mydir/theano-cuda-activate.sh && ln -s $mydir/theano-cuda-activate.sh $train_dir/
 for f in [
         "data-train", "tools", "tools-recog",
         "flow", "features", "features.warped",
@@ -48,16 +81,6 @@ for f in [
         "sprint-executables"]:
     if os.path.exists("%s/%s" % (mydir, f)):
         os.symlink("base/%s" % f, "%s/%s" % (train_dir, f))
-
-# Doesnt work for now. Maybe git-new-workdir + submodule combination is broken?
-
-# Create CRNN branch which we use for the setup.
-#cd crnn
-#git branch $model -f
-#cd ..
-
-#echo "git-new-workdir crnn"
-#git-new-workdir crnn $train_dir/crnn $model
 
 # Fall-back:
 os.symlink("base/%s" % Settings.returnn_dir_name, "%s/%s" % (train_dir, Settings.returnn_dir_name))
